Use Pundit where it's perfunctory
In this PR https://github.com/Cammi-Technology/sparkling-water/pull/1 we
discuss wether skipping the autorization check in pundit should be used
in places where Pundit will always return true. With arguments for
consistency, vs against additional scaffolding.

This commit implements consistency in an example where we've also argued
that we _should_ skip the check where (and only where) authorization is
not a concern at all.

Discuss!