Use specified policy class
In 636edcdbc8e87b3917c03b99239be4b73f8b0d74 we look at using Pundit
where it may not be strictly necessary, on the PR I commented that I am
not keen on specifying a custom action to be checked for authorization.

https://github.com/Cammi-Technology/sparkling-water/pull/1/commits/636edcdbc8e87b3917c03b99239be4b73f8b0d74#r2124053369

Which you prefer may come down to if a Policy maps a model, or a
controller in your mind. This commit works on the assumption that a
Policy maps the controller.

Arguably this is false, given Policies are inferred by the resource