Add pundit to identity controller actions
The approach I have gone for is to decide that a Policy class will match
the resource passed into it. Other approaches were discussed in #1 such
as making a policy for the controller rather than the resource. In the
end I didn't like that since it disconnected a policy from its resource.