Update policies
There has been a lot of discussion about policies, the last agreement we
made was to remove user policies and only have account user policies.

This will eventually be impractical, as we will definetly want to
control who can create accounts, probably based on if they're paying or
not.

I'm kind of into the idea of the policiy using an array of account users
for the current user. Within an account this would be scoped to their
current account. Outside of an account, it would be all of them. That
way we'd always have access to the user, but guess what, we'd also
always have access to account user too! — maybe!

For now I want to move onto something more fun.