Add syntax highlighting for Post & Comment bodies
While we wait for a swanky new text editor from 37Signals which likely
does syntax highlighting for us we can implement something short-term,
simple and easy to remove: frontend-only syntax highlighting with
highlight.js.

Highlight.js is neat as it auto-detects languages for us (so we don't
need to extend trix with a language selection dropdown), and supports a
huge range of languages.

If we want a different theme we can replace `syntax_highlighting.css`
with a highlight.js theme in the future. Currently we only support Ruby,
Python, Javascript, HTML and CSS - but highlightjs supports lots more.

The pins for highlightjs may look quite funny, this is a workaround for
a rails and/or highlightjs importmaps issue <https://github.com/rails/importmap-rails/issues/270>